Create NQueen.cpp
For the heuristic i decided to go with a mine sweeper kind of tactic as the program goes through the columns it finds all the queens possible positions it then creates a 3x3 matrix around that queen and takes the sum of all the nodes. All the possible positions are paired with their weight and are put into a min heap priority queue. the find next function then takes the best option and back tracks on to the next best if that one doesn't work.